Bootcamps/Bradesco - GenAI & Dados/Desafio de Projeto Final — Construa seu Assistente Virtual com IA Generativa/requirements.txt
requests
python-dotenv
google-generativeai
brapi